aries-common: Support custom mkbootimg
Currently, we build both the recovery ramdisk and the main ramdisk into the kernel.  This done with the flag BOARD_CUSTOM_BOOTIMG_MK, which is not an AOSP-supported flag (CAF originally?).

Instead, lets store the ramdisks on our previously unused (except in the case of fascinatemtd - not sure if the reason is still valid though) second mtd partition.  This has been relabelled to "ramdisk" (from "recovery").  Our stage1 ramdisk now mounts that partitions and loads the correct ramdisk.

The other reason BOARD_CUSTOM_BOOTIMG_MK was used was so that we didn't append the ramdisk afterwards.  Now, we use a BOARD_CUSTOM_MKBOOTIMG which simply copies the kernel to the boot.img (or whatever the output flag is).

This has some advantages and disadvantages to the previous method:
Advantages:
-The size of the boot.img is reduced as the ramdisks are stored externally
-We now use the previously unused partition on the onenand
-Is more AOSP-compatible as the remaining function of BOARD_CUSTOM_BOOTIMG_MK is solely to build the busybox and erase_image executables for inline kernel building (which AOSP doesn't support)
-Is easier to manually edit the ramdisks (before it was essentially impossible)
-We can independently update the recovery ramdisk from the ROM
-Could easily be modified for multi-ROM kernels (ie add another ramdisk for CM)...
Disadvantages:
-Is a little bit more cumbersome in the updater.sh and the releasetools
-Doesn't support the fascinatemtd (which could be changed fairly easily, presuming that the recovery_kernel that is currently flashed is indeed unused...)
-Still not totally AOSP-compatible, but this can be easily changed as listed above.
-Slightly longer bootup time

I have also increased the colboot timeout in this patch (we were over the limit before anyways).

You can still flash a "regular" kernel overtop and it will use its built-in ramdisk's without a problem

Change-Id: If561631c3084ac905acdeeb3e0846371d6a41e99